+++ src/components/outside_click_detector/index.ts
@@ -31,4 +31,5 @@
 export {
   OuiOutsideClickDetector,
   OuiOutsideClickDetectorProps,
+  OuiEvent,
 } from './outside_click_detector';